% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{determine_p0}
\alias{determine_p0}
\title{Assistance function for ranking}
\usage{
determine_p0(mu, n)
}
\arguments{
\item{mu}{Multiplicative constant}

\item{n}{Number of items to rank}
}
\description{
p0 is defined to be a constant for which the raking program builds off of.
sum_(i = 1)^(n) p_0 * (mu)^i = 1
}
\examples{
determine_p0(.65,10)

}
\keyword{p0}
\keyword{rankr}

