% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{gg_colors}
\alias{gg_colors}
\title{ggplot2 Color Defaults}
\usage{
gg_colors(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
The colors from the great ggplot2 package by Hadley Wickhim which are
evenly spaced hues aroudn the hcl colour wheel
}
\examples{
scales::show_col(gg_colors(4))
}
\keyword{colors}
\keyword{ggplot2}

