% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{Google_Color_Palette}
\alias{Google_Color_Palette}
\title{For this file, the following is useful: library(roxygen2) library(devtools) 
From the package parent directory: Create a package with:
devtools::create("packagename") You can document it with
devtools::document("packagename") and documenting it allows the function to
be visible Install it with devtools::install("packagename") which will make
it usable to all. 
Download from github with: devtools::install_github("username/projectName")}
\usage{
Google_Color_Palette(n = 20)
}
\arguments{
\item{n}{The number of colors to select for the palette}
}
\description{
For this file, the following is useful: library(roxygen2) library(devtools) 
From the package parent directory: Create a package with:
devtools::create("packagename") You can document it with
devtools::document("packagename") and documenting it allows the function to
be visible Install it with devtools::install("packagename") which will make
it usable to all. 
Download from github with: devtools::install_github("username/projectName")
}
\examples{
scales::show_col(Google_Color_Palette())
}
\keyword{Google}
\keyword{colors}

