% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{Google_Logo_Palette}
\alias{Google_Logo_Palette}
\title{Google Logo Palette}
\usage{
Google_Logo_Palette()
}
\arguments{
\item{none}{none}
}
\description{
This function allows you to use all of the colors that are common in a google
logo charts implementation and can be used with the palette command to set
the colors
}
\examples{
scales::show_col(Google_Logo_Palette())
}
\keyword{Google}
\keyword{colors}

