% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{huber_loss}
\alias{huber_loss}
\title{Huber Loss Function}
\usage{
huber_loss(x, delta = 1)
}
\arguments{
\item{x}{residual value}

\item{delta}{cutoff point for the loss function}
}
\description{
The huber loss function is quadratic for x < delta 
and then linear past this
}
\examples{
huber_loss(1)
}
\keyword{huber}
\keyword{loss}

