% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{mod_huber_loss}
\alias{mod_huber_loss}
\title{Modified Huber Loss Function}
\usage{
mod_huber_loss(x, delta = 1)
}
\arguments{
\item{x}{residual value}

\item{delta}{cutoff point for the loss function}
}
\description{
The modified huber loss function acts like the traditional huber loss
function but is analytic and has all derivatives cont.
}
\examples{
mod_huber_loss(1)
}
\keyword{huber}
\keyword{loss}

