% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{plot_rank_2d_contour}
\alias{plot_rank_2d_contour}
\title{Visualization of the Rank Space}
\usage{
plot_rank_2d_contour(mu, n = 20, t = 2, weights = 1, plotContour = TRUE)
}
\arguments{
\item{mu}{Multiplicative constant}

\item{n}{Number of items to rank}

\item{t}{Power that controls degree of curvature of the Goodness of Fit lines}
}
\description{
A plotting function which tries to give an intuitive view of how things are
ranked with different parameters
}
\examples{
plot_rank_2d_contour(mu = 0.25, n = 15)
}
\keyword{plot}
\keyword{rankr}

