% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankr_main.R
\name{rankr}
\alias{rankr}
\title{Multifactor Ranking}
\usage{
rankr(data, highlow = NULL, targets = NULL, mu = 0.99, t = 2,
  return_weights = F, weights = NULL, display_chart = F,
  bar_n_display = 10, text_n_display = 30)
}
\arguments{
\item{data}{Data frame to rank}

\item{highlow}{vector of "H", "L", "T" that indicates how to rank columns}

\item{targets}{vector of NA's and targets to hit, if applicable}

\item{mu}{is the geometric decreasing parameter}

\item{t}{is the norm calculation power parameter}
}
\description{
A plotting function which tries to give an intuitive view of how things are
ranked with different parameters
}
\examples{
# Find the flowers in iris with the Highest Sepal Length and Lowest width
mydata = iris[1:2]
ranks <- rankr(mydata, highlow = c("H", "L"), targets = rep(NA,2))
plot(mydata[ranks, ], type = "n", xlim = c(4,8), ylim = c(1,5))
text(mydata[ranks, ], labels = seq_along(ranks))
}
\keyword{multifactor}
\keyword{plot}
\keyword{ranking}
\keyword{rankr}

