% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{table_to_vectors}
\alias{table_to_vectors}
\title{Convert Table Object to Vectors}
\usage{
table_to_vectors(tab)
}
\arguments{
\item{tab}{A table object.}
}
\value{
A list of two vectors `x, y` satisfying `table(x, y) == tab`.
}
\description{
Convert Table Object to Vectors
}
