% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tau}
\alias{tau}
\alias{tau_a}
\alias{tau_b}
\alias{tau_c}
\alias{tau_d}
\alias{somers_d}
\alias{gk_gamma}
\title{Estimate variants of Kendall's tau}
\usage{
tau_a(x, y)

tau_b(x, y)

tau_c(x, y)

tau_d(x, y)

somers_d(x, y)

gk_gamma(x, y)
}
\arguments{
\item{x, y}{Vectors of observations.}
}
\value{
The variant of Kendall's tau.
}
\description{
Estimate variants of Kendall's tau
}
