% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\alias{tau_alt_weak}
\alias{tau_alt_stric}
\alias{tau_strict}
\alias{tau_a}
\alias{tau_b}
\alias{tau_c}
\alias{somers_d}
\alias{wilsons_e}
\alias{gk_gamma}
\alias{lg_d}
\title{Estimate variants of Kendall's tau for tied data}
\usage{
tau_alt_weak(x, y)

tau_alt_stric(x, y)

tau(x, y)

tau_strict(x, y)

tau_a(x, y)

tau_b(x, y)

tau_c(x, y)

somers_d(x, y)

wilsons_e(x, y)

gk_gamma(x, y)

lg_d(x, y)
}
\arguments{
\item{x}{Covariate data.}

\item{y}{Response data.}
}
\value{
Kendall's tau.
}
\description{
Estimate variants of Kendall's tau for tied data
}
\section{Functions}{
\itemize{
\item \code{tau_alt_weak()}: Alternative weakly monotone PPR tau (Moss, 2024)

\item \code{tau_alt_stric()}: Alternative strictly monotone PPR tau (Moss, 2024)

\item \code{tau()}: Weakly monotone PPR tau (Moss, 2024)

\item \code{tau_strict()}: Strictly monotone PPR tau (Moss, 2024)

\item \code{tau_a()}: Original Kendall's tau (Kendall, 1938)

\item \code{tau_b()}: Kendall's tau b (Kendall, 1945)

\item \code{tau_c()}: Stuart's tau (Stuart, 1953)

\item \code{somers_d()}: Somers' D (Somers, 1962)

\item \code{wilsons_e()}: Wilson's e (Wilson, 1974)

\item \code{gk_gamma()}: Goodman-Kruskall gamma (Goodman, Kruskall, 1954)

\item \code{lg_d()}: Leik-Gove D (Leik, Gove, 1969)

}}
